% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{accuracy}
\alias{accuracy}
\title{Vector Comparison Function}
\usage{
accuracy(probabilities, outcomes, cutoff = 0.5)
}
\arguments{
\item{probabilities}{Vector of probabilities to be analyzed. Required parameter.}

\item{outcomes}{Vector of outcomes used to grade the vector of probabilities. Required parameter}

\item{cutoff}{Threshold used to determine point at which probabilities are successes. Defaults to 0.5.}
}
\description{
This function allows you to compare vectors.
}
\examples{
accuracy()
}
\keyword{accuracy}
